Move monitor into dockerAPI
