+++ supervisor/coresys.py
@@ -13,7 +13,6 @@ import sentry_sdk
 
 from .config import CoreConfig
 from .const import ENV_SUPERVISOR_DEV, SERVER_SOFTWARE
-from .docker import DockerAPI
 from .utils.dt import UTC, get_time_zone
 
 if TYPE_CHECKING:
@@ -26,6 +25,7 @@ if TYPE_CHECKING:
     from .core import Core
     from .dbus.manager import DBusManager
     from .discovery import Discovery
+    from .docker.manager import DockerAPI
     from .docker.monitor import DockerMonitor
     from .hardware.manager import HardwareManager
     from .homeassistant.module import HomeAssistant
@@ -64,9 +64,9 @@ class CoreSys:
 
         # Global objects
         self._config: CoreConfig = CoreConfig()
-        self._docker: DockerAPI = DockerAPI()
 
         # Internal objects pointers
+        self._docker: DockerAPI | None = None
         self._core: Core | None = None
         self._arch: CpuArch | None = None
         self._auth: Auth | None = None
